Add original Efile data for Form 700
Based on user testing, we learned that our it would be valuable to give
a better overview of Form 700s in the email so that users don't have to
click through for people they marginally care about. Instead, they can
just skim through the email for these people.

This downloads the "Efile.txt", an XML document, from Netfile for the
Form 700's. In order to hold the logic for processing the raw data
returned from Netfile for this form, as well as for Form 460, I've
introduced a new abstraction in app/lib/forms where those helper classes
are responsible for calculating the important numbers based off the raw
form contents.