Add human-friendly descriptions of forms
Although the email is targeted at experts right now, let's see if we can
make it a bit more inclusive without adding too much extra content.

This commit adds section headings for each type of form, groups the
forms by type, and, at the beginning of the section heading, includes a
bit of introductory material to the form.

Maybe, eventually, we can actually keep track of the state of onboarding
for each user, and skip the descriptions if the user's seen it a million
times already.