Add associations to get the spreadsheet models
Previous commits import the spreadsheet into tables starting with
`election`. By creating associations from Filing, we can use the
filing's `filer_id` column to fetch the related objects.

Candidates and Committees are easy, we can simply match the filing's
filer_id with the value of "Filer_ID" on the spreadsheet. Referendums
are trickier, we have to go from Filing to Committee to Referendum
according to the "Ballot Measure" and "Ballot Measure Election" columns.