Combine multiple 496's from same filer
This is the first attempt at editorializing the data somewhat: due to
filing quirks, it appears that Independent Expenditure committees
supporting or opposing multiple candidates typically file a separate
496 for each candidate supported or opposed. This clogs up the alert
email with largely repetitive and sometimes misleading information
(because the contributions are typically reported on all 496 forms).

Forms will be combined for display if they are from the same filer, and
both have either no contributions, or at least one contribution in
common with each other.

In the future this functionality will be used to remove multiple
amendments from the same update email as well (#20).

Closes #41.