Add confirmation step to prevent spam
In the last week we've had about 250 almost-certainly spam accounts
subscribe. They don't even open the welcome email -- how rude!

This commit adds a `confirmed_at` timestamp that's populated when the
user clicks a link in the initial email. Until confirmed, an
AlertSubscriber is not considered `subscribed` for the purposes of daily
sending or the default admin UI.

In production, I will be auto-confirming all users that were subscribed
before the spam started (roughly November 7) and I will send out
manual confirmation emails to all the suspicious subscribers after that.