Fix crash when combining amended Form 496/497's with non-amended
There was a bug where the combination of one amended filing (from
yesterday's 497) with today's unanmended filing would result in an error
because the constructor for the combined form class would be called with
a singular filing by the base class.

To avoid that, let's avoid combining amended returns with non-amended
returns. This seems like it will be good UX, to me, since the point of
showing the amended forms in the email is to highlight the change versus
the original. If we were to combine amended forms with other form
originals, then it may confuse people as they see double transactions.